Update based on comments
