+++ lib/avro_turf/schema_store.rb
@@ -1,5 +1,4 @@
 class AvroTurf::SchemaStore
-  attr_accessor :schemas
 
   def initialize(path: nil)
     @path = path or raise "Please specify a schema path"
@@ -58,12 +57,4 @@ class AvroTurf::SchemaStore
     end
   end
 
-  # @param schema_hash [Hash]
-  def add_schema(schema_hash)
-    name = schema_hash['name']
-    namespace = schema_hash['namespace']
-    full_name = Avro::Name.make_fullname(name, namespace)
-    return if @schemas.key?(full_name)
-    Avro::Schema.real_parse(schema_hash, @schemas)
-  end
 end
